feat: move storm tracking backend from ESRC to EvE-Scout API
This PR refactors the storm tracking feature so data is no longer stored in the ESRC database.
Instead, updates send an API request to the EvE-Scout API, where the data gets stored and provided to the greater EVE community.
Moreover, we are now tracking the Space Oddity "Tom's Shuttle" as well.